--- client/html/splash/index.html
@@ -1,18 +1,0 @@
-<html>
-    <body
-        style="font-family: Arial; background-color: transparent; text-align: center; color: white"
-    >
-        <h3 style="margin-top: 50vh">
-            Powered by Open Roleplay
-        </h3>
-        <ul style="list-style-type: none; padding: 0; margin: 0;">
-            <li>https://twitch.tv/stuykgaming/</li>
-            <li>https://github.com/team-stuyk-alt-v/altV-Open-Roleplay</li>
-        </ul>
-        <script
-            src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min.js"
-            type="text/javascript"
-        ></script>
-        <script src="./app.js"></script>
-    </body>
-</html>
