+++ client/locations/locations.mjs
@@ -1,4 +1,4 @@
-export const Locations = [
+export const Atms = [
     {
         x: -1109.797,
         y: -1690.808,
@@ -242,21 +242,21 @@ export const Locations = [
     {
         x: 146.769,
         y: -1035.389,
-        z: 29.330
+        z: 29.33
     },
     {
-        x: 119.090,
+        x: 119.09,
         y: -883.753,
         z: 31.116
     },
     {
-        x: -256.140,
+        x: -256.14,
         y: -715.872,
         z: 33.509
     },
     {
         x: -258.843,
-        y: -723.560,
+        y: -723.56,
         z: 33.458
     },
     {
@@ -272,17 +272,17 @@ export const Locations = [
     {
         x: 111.309,
         y: -775.371,
-        z: 31.436   
+        z: 31.436
     },
     {
-        x: 114.408, 
+        x: 114.408,
         y: -776.373,
-        z: 31.402   
+        z: 31.402
     },
     {
-        x: 285.613, 
+        x: 285.613,
         y: 143.406,
-        z: 104.160   
+        z: 104.16
     },
     {
         x: -56.887,
@@ -301,7 +301,70 @@ export const Locations = [
     },
     {
         x: -29.169,
-        y: -723.890,
+        y: -723.89,
         z: 44.225
+    }
+];
+
+export const Clothing = [
+    {
+        x: -160.74725341796875,
+        y: -302.8483581542969,
+        z: 39.7266845703125
+    },
+    {
+        x: -1107.5208740234375,
+        y: 2709.81103515625,
+        z: 19.1025390625
+    },
+    {
+        x: -3172.510009765625,
+        y: 1048.0087890625,
+        z: 20.85498046875
+    },
+    {
+        x: -1193.3670654296875,
+        y: -772.3384399414062,
+        z: 17.3165283203125
+    },
+    {
+        x: 618.03955078125,
+        y: 2759.406494140625,
+        z: 42.085693359375
+    },
+    {
+        x: 123.5868148803711,
+        y: -219.40219116210938,
+        z: 54.5545654296875
+    },
+    {
+        x: 11.024175643920898,
+        y: 6512.74267578125,
+        z: 31.874755859375
+    },
+    {
+        x: -1453.041748046875,
+        y: -236.36044311523438,
+        z: 49.802978515625
+    },
+    {
+        x: -828.5406494140625,
+        y: -1072.879150390625,
+        z: 11.3179931640625
+    },
+    {
+        x: 428.9142761230469,
+        y: -801.112060546875,
+        z: 29.4820556640625
+    },
+    {
+        x: 71.57801818847656,
+        y: -1397.96044921875,
+        z: 29.3641357421875
+    },
+    {
+        x: -710.5977783203125,
+        y: -155.2087860107422,
+        z: 37.4014892578125
     }
 ];
