+++ client/panels/character.mjs
