+++ client/panels/clothing.mjs
