+++ client/systems/character.mjs
@@ -1,7 +1,7 @@
 import * as alt from 'alt';
 import * as native from 'natives';
 
-alt.log('Loaded: client->character->face.mjs');
+alt.log('Loaded: client->systems->character.mjs');
 
 // Used for facial features loop.
 const faceFeatureNames = [
@@ -26,6 +26,35 @@ const faceFeatureNames = [
     'ChinShape',
     'NeckWidth'
 ];
+
+// Synchronize the clothing sent down from the server.
+export function syncClothing(jsonData) {
+    const data = JSON.parse(jsonData);
+
+    for (let key in data) {
+        if (!data[key].isProp) {
+            native.setPedComponentVariation(
+                alt.Player.local.scriptID,
+                data[key].id,
+                data[key].value,
+                data[key].texture,
+                0
+            );
+        } else {
+            native.setPedPropIndex(
+                alt.Player.local.scriptID,
+                data[key].id,
+                data[key].value,
+                data[key].texture,
+                true
+            );
+
+            if (data[key].value === -1) {
+                native.clearPedProp(alt.Player.local.scriptID, data[key].id);
+            }
+        }
+    }
+}
 
 // Apply all of the facial data to the player.
 export function applyFacialData(jsonData) {
