--- server/serverEvents/events.mjs
@@ -1,5 +1,0 @@
-// On load; send a message.
-console.log('Loaded: serverEvents->serverEventRouting.mjs');
-
-// This file is meant as a way to organize server events and import accordingly.
-//alt.on('chatIntercept', chatChat.relayChat);
